RAN 725B

+++ Problem725B.cpp
@@ -60,7 +60,7 @@ int main() {
     string s;
     cin >> s;
     int t = 0;
-    int k = s[0] - '0';
+    ll k = s[0] - '0';
     if(s[1] == 'f') {
         t += 1;
     } else if(s[1] == 'e') {
